Create Chapter 3-7
In cold weather, meteorologists report an index called the wind chill factor, which
takes into account the wind speed and the temperature. The index provides a
measure of the chilling effect of wind at a given air temperature. Wind chill may
be approximated by the following formula:
w = 33 - (102v - v + 10.5)(33 - t)
23.1
 where
v = wind speed in m/sec
t = temperature in degrees Celsius: t 6= 10
W = wind chill index (in degrees Celsius)
 Write a function that returns the wind chill index. Your code should ensure that
the restriction on the temperature is not violated. Look up some weather reports in
back issues of a newspaper or from a weather website and compare the wind chill
index you calculate with the result reported or website.