Create Chapter 5-4
Write a program that reads in an array of type int. You may assume that there are
fewer than 50 entries in the array. Your program determines how many entries are
used. The output is to be a two-column list. The first column is a list of the distinct
array elements; the second column is the count of the number of occurrences of
each element. The list should be sorted on entries in the first column, largest to
smallest.
 For the array values
–12 3 −12 4 1 1 −12 1 −1 1 2 3 4 2 3 −12
 the output should be
 N Count
 4 2
 3 3
 2 2
 1 4
–1 1
–12 4